+++ client/src/pages/write/index.tsx
@@ -11,7 +11,7 @@ import * as Yup from 'yup';
 import { errorMessage } from '@/constant';
 
 //service
-import { boardPost } from '@/service/BoardPost';
+import { boardPost } from '@/service/boardPost';
 import { Icon } from '@iconify/react';
 
 // TODO : title 없을 시 빨간 테두리
